[#71] 파일 분리 및 디렉터리 정리
